clean code and add javadoc
Closes #57
